pyproject.toml
[project]
name = "mails-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi[standard]>=0.115.8",
    "google-api-python-client>=2.161.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "itsdangerous>=2.2.0",
    "openai>=1.63.2",
    "python-dotenv>=1.0.1",
]
